refactor(extractor): moving files
