+++ src/scripts/content/extractor/rules/rules.js
@@ -1,4 +1,4 @@
-import {turndownConverter} from "../../shared/formatter/formatMarkdown";
+import {turndownConverter} from "../../../shared/formatter/formatMarkdown";
 
 export const rules = {
   'reformatLinksContainingTags': {
